Add database serialization and remove old json module
