+++ stack-graphs/tests/it/main.rs
@@ -21,8 +21,6 @@ mod can_jump_to_definition;
 mod can_jump_to_definition_with_forward_partial_path_stitching;
 mod cycles;
 mod graph;
-#[cfg(feature = "json")]
-mod json;
 mod partial;
 #[cfg(feature = "serde")]
 mod serde;
